changing length and block size to long long

+++ aes.h
@@ -54,7 +54,7 @@ void AES_ctx_set_iv(struct AES_ctx* ctx, const uint8_t* iv);
 // Suggesting https://en.wikipedia.org/wiki/Padding_(cryptography)#PKCS7 for padding scheme
 // NOTES: you need to set IV in ctx with AES_init_ctx_iv() or AES_ctx_set_iv()
 //        no IV should ever be reused with the same key 
-void AES_CTR_xcrypt_buffer(struct AES_ctx* ctx, uint8_t* buf, uint32_t length, int my_rank, int comm_sz, MPI_Comm comm);
+void AES_CTR_xcrypt_buffer(struct AES_ctx* ctx, uint8_t* buf, long long int length, int my_rank, int comm_sz, MPI_Comm comm);
 
 #endif // #if defined(CTR) && (CTR == 1)
